Eligibility Service conversion to Python
The eligibility has been rewritten in full Python. There is a SQL Server specific method MonthsAdd that was necessary. When migrating to PostgreSQL, one can replace it with a `datetimedelta(month=1) * F("waitingperiod")`, that doesn't work in SQL Server.
Both services are currently maintained and the tests run both to compare the results.

The service also exposes two signals: signal_eligibility_service_before and signal_eligibility_service_after. They pass along the request and the response. Setting the response.final to True will stop processing further and use that response.
It is also possible to disable the provided eligibility service with `default_eligibility_disabled`.

For testing convenience, I have changed the test helper for Policy creation to return a tuple with Policy and InsureePolicy. It is compatible with the original method but using the `create_test_policy2` method is easier.