List and delete policy renewals with gql or when renewing policy
Policy renewals were directly exposed by the C# REST API. The policy mobile app
needs to list, delete, and add policy renewals. The GraphQL API didn't allow
to do so.

With the present change, we can now list and delete policy renewals with a
GrapQL query and mutation (resp.). As there is already a mutation to renew
a given policy, we haven't added a dedicated GraphQL mutation to add
a policy renewal independently. As there is a background task scheduling policy
renewals, we are now soft deleting the renewals of a given policy when we
manually renew it.